RIA-5293 refactor & test coverage
